// src/types.ts
export enum GraduationStatus {
  LULUS = "LULUS",
  TIDAK_LULUS = "TIDAK LULUS"
}

export interface Student {
  id?: string;
  nama: string;
  nisn: string;
  status: GraduationStatus;
  certificateUrl?: string;
  updatedAt?: any;
}

export enum OperationType {
  CREATE = 'create',
  UPDATE = 'update',
  DELETE = 'delete',
  LIST = 'list',
  GET = 'get',
  WRITE = 'write',
}

export interface FirestoreErrorInfo {
  error: string;
  operationType: OperationType;
  path: string | null;
  authInfo: {
    userId?: string | null;
    email?: string | null;
    emailVerified?: boolean | null;
    isAnonymous?: boolean | null;
    tenantId?: string | null;
    providerInfo?: {
      providerId?: string | null;
      email?: string | null;
    }[];
  }
}